feat: switch from simulation to live trading mode
- Fix BrokerOrder.quantity from int to float for fractional crypto amounts
- Auto-connect trader to MEXC exchange on app startup
- Auto-start trading loop if enabled in config (LIVE mode)
- Add live_mode/mode fields to trading status API
- Show LIVE badge with pulse animation in UI when connected
- Clear visual distinction between LIVE and disconnected states

+++ brokers/base.py
@@ -23,7 +23,7 @@ class OrderType(Enum):
 class BrokerOrder:
     ticker: str
     side: OrderSide
-    quantity: int
+    quantity: float
     order_type: OrderType = OrderType.MARKET
     limit_price: Optional[float] = None
     order_id: Optional[str] = None
@@ -109,7 +109,7 @@ class BaseBroker(ABC):
         """Annule un ordre. Retourne True si succès."""
         pass
 
-    def buy(self, ticker: str, quantity: int, order_type: OrderType = OrderType.MARKET,
+    def buy(self, ticker: str, quantity: float, order_type: OrderType = OrderType.MARKET,
             limit_price: float = None) -> BrokerOrder:
         """Raccourci pour placer un ordre d'achat."""
         order = BrokerOrder(
@@ -121,7 +121,7 @@ class BaseBroker(ABC):
         )
         return self.place_order(order)
 
-    def sell(self, ticker: str, quantity: int, order_type: OrderType = OrderType.MARKET,
+    def sell(self, ticker: str, quantity: float, order_type: OrderType = OrderType.MARKET,
              limit_price: float = None) -> BrokerOrder:
         """Raccourci pour placer un ordre de vente."""
         order = BrokerOrder(
